Remove many unnecessary dependencies
I noticed that the version of this repo that @stevector modified for
Drupal 9 removed a lot of the dependencies, in particular require-dev.
This makes sense to me because
 * This repo doesn't have anything like a particular CI workflow we
encourage folks to use that actually makes use of these dependencies
 * The more dependencies we start off with, the more things we/custom
upstream maintainers that fork this need to keep an eye on over time
(unsupported major versions etc.)

This would make the drupal 8 repo consistent with the drupal 9 one in
terms of not starting off with these things.